backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
docx2txt
